+++ gridstatus/caiso.py
@@ -100,8 +100,7 @@ class CAISO(ISOBase):
         return self._get_historical_fuel_mix(date, verbose=verbose)
 
     def _get_historical_fuel_mix(self, date, verbose=False):
-        url = _HISTORY_BASE + "/%s/fuelsource.csv"
-        df = _get_historical(url, date, verbose=verbose)
+        df = _get_historical("fuelsource", date, verbose=verbose)
 
         # rename some inconsistent columns names to standardize across dates
         df = df.rename(
@@ -137,8 +136,8 @@ class CAISO(ISOBase):
         return self._get_historical_load(date, verbose=verbose)
 
     def _get_historical_load(self, date, verbose=False):
-        url = _HISTORY_BASE + "/%s/demand.csv"
-        df = _get_historical(url, date, verbose=verbose)
+        df = _get_historical("demand", date, verbose=verbose)
+
         df = df[["Time", "Current demand"]]
         df = df.rename(columns={"Current demand": "Load"})
         df = df.dropna(subset=["Load"])
@@ -335,17 +334,16 @@ class CAISO(ISOBase):
         if date == "latest":
             return self._latest_from_today(self.get_storage)
 
-        url = _HISTORY_BASE + "/%s/storage.csv"
-        df = _get_historical(url, date, verbose=verbose)
-        # new column in data
-        # ['Time', 'Total batteries', 'Stand-alone batteries', 'Hybrid batteries',
-        # 'Type']
-        if "Total batteries" in df.columns:
-            df = df.rename(columns={"Total batteries": "Supply"})
-            df = df.drop(columns=["Stand-alone batteries", "Hybrid batteries"])
-        else:
-            df = df.rename(columns={"Batteries": "Supply"})
-        df["Type"] = "Batteries"
+        df = _get_historical("storage", date, verbose=verbose)
+
+        df = df.rename(
+            columns={
+                "Total batteries": "Supply",
+                "Stand-alone batteries": "Stand-alone Batteries",
+                "Hybrid batteries": "Hybrid Batteries",
+            },
+        )
+        df = df[["Time", "Supply", "Stand-alone Batteries", "Hybrid Batteries"]]
         return df
 
     @support_date_range(frequency="31D")
@@ -814,13 +812,18 @@ def _make_timestamp(time_str, today, timezone="US/Pacific"):
     )
 
 
-def _get_historical(url, date, verbose=False):
-    date_str = date.strftime("%Y%m%d")
-    date_obj = date
-    url = url % date_str
-
-    msg = f"Fetching URL: {url}"
-    log(msg, verbose)
+def _get_historical(file, date, verbose=False):
+    try:
+        date_str = date.strftime("%Y%m%d")
+        url = _HISTORY_BASE + "/%s/%s.csv" % (date_str, file)
+        if verbose:
+            print("Fetching URL: ", url)
+    except Exception:
+        # fallback if today and no historical file yet
+        if utils.is_today(date, CAISO.default_timezone):
+            url = _BASE + "/%s.csv" % file
+            if verbose:
+                print("Fetching URL: ", url)
 
     df = pd.read_csv(url)
 
@@ -829,7 +832,7 @@ def _get_historical(url, date, verbose=False):
 
     df["Time"] = df["Time"].apply(
         _make_timestamp,
-        today=date_obj,
+        today=date,
         timezone="US/Pacific",
     )