+++ gridstatus/ercot.py
@@ -165,11 +165,11 @@ class Ercot(ISOBase):
             r = self._get_json(url, verbose=verbose)
 
             today_str = date.strftime("%Y-%m-%d")
-
             mix = (
                 pd.DataFrame(r["data"][today_str])
                 .applymap(
                     lambda x: x["gen"],
+                    na_action="ignore",
                 )
                 .T
             )
@@ -286,7 +286,7 @@ class Ercot(ISOBase):
 
         Currently only supports today's forecast
         """
-        if date != "today":
+        if not utils.is_today(date, self.default_timezone):
             raise NotSupported()
 
         # intrahour https://www.ercot.com/mp/data-products/data-product-details?id=NP3-562-CD