+++ gridstatus/tests/test_version.py
@@ -2,4 +2,4 @@ from gridstatus import __version__
 
 
 def test_version():
-    assert __version__ == "0.18.0"
+    assert __version__ == "0.19.0"