+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## vNext
 
 - Updated ISONE Interconnection Queue to contain completed and withdrawn projects
+- Add all areas to PJM get_load
 
 ## v0.18.0 - Jan 27, 2023