+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
-## vNext
+## v0.19.0 - Feb 19, 2023
 
 - Updated ISONE Interconnection Queue to contain completed and withdrawn projects
+- Add all areas to PJM get_load
+- Add load over time visualization
 
 ## v0.18.0 - Jan 27, 2023